Dodanie klas zapisujących do pliku. Uzupełnienie komentarzy. Usunięcie kilku wycieków pamięci, zmiana nazw kilku zmiennych i fukcji na bardziej intyicyjne.

+++ PredykcjaCiagu/Logic/NeuralNetworkComponents/Neuron.h
@@ -89,13 +89,13 @@ public:
 	}
 
 	/** Getter do pojedynczej wagi neurona */
-	double GetOneWeight(double whichInput)
+	double GetOneWeight(int whichInput)
 	{
 		return weightsOfNeuronInputs[whichInput];
 	}
 
 	/** Setter do pojedynczej wagi neurona */
-	void SetOneWeight(double whichInput, double tWeight)
+	void SetOneWeight(int whichInput, double tWeight)
 	{
 		weightsOfNeuronInputs[whichInput] = tWeight;
 	}
